HTML Modularizado, carpetas nuevas de cada parte
