+++ src/components/Footer/AtencionLocal/AtencionLocal.jsx
@@ -1,0 +1,37 @@
+import React from 'react'
+
+function AtencionLocal() {
+  return (
+    <section className="AtencionLocal">
+              <address>
+                <p>Ubicacion</p>
+                <p>Distrito Numero1</p>
+                <p>Calle Las Florecitas</p>
+              </address>
+            <div>
+                <h4>Atención:</h4>
+                <p>Lunes-Viernes</p>
+                <p>12:00 am- 5:00 pm</p>
+                <p>Sábado</p>
+                <p>10:00 am - 6:00 pm</p>
+            </div>
+            <div>
+              <ul>
+                <li><a href="">Home</a></li>
+                <li><a href="">Recetario</a></li>
+                <li><a href="">Pasteles</a></li>
+                <li><a href="">Nosotros</a></li>
+              </ul>
+            </div>
+            <div>
+              <ul>
+                <li>Club Galletas</li>
+                <li>Club Pasteles</li>
+                <li>Club Especial</li>
+              </ul>
+            </div>
+        </section>
+  )
+}
+
+export default AtencionLocal
