+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import LastSection from './LastSection/LastSection'
+import AtencionLocal from './AtencionLocal/AtencionLocal'
+
+function Footer() {
+  return (
+    <footer>
+        <LastSection/>
+        <AtencionLocal/>
+    </footer>
+  )
+}
+
+export default Footer
