+++ src/components/Footer/LastSection/LastSection.jsx
@@ -1,0 +1,23 @@
+import React from 'react'
+
+function LastSection() {
+  return (
+    <section className='LastSection'>
+          <img src="" alt="" />
+          <h2></h2>
+          <div className="Redes">
+            <div><img src="" alt="" />
+            
+            </div>
+            <div><img src="" alt="" />
+            
+            </div>
+            <div><img src="" alt="" />
+            
+            </div>
+          </div>
+        </section>
+  )
+}
+
+export default LastSection
