+++ src/components/Main/Milkshakes/Milkshakes.jsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+function Milkshakes() {
+  return (
+    <section className='Milshakes'>
+        <img src="" alt="Imagen de milkshake" />
+        <div className='CouruselMilshake'>
+        <h3>¿Con qué los tomamos?</h3>
+        <div>
+            <h5>Fruity</h5>
+            <img src="" alt="Imagen de milkshake frutal" />
+        </div>
+        </div>
+    </section>
+  )
+}
+
+export default Milkshakes
