+++ src/components/Main/Process/Process.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+function Process() {
+  return (
+    <section className='Process'>
+          <h3>Nuestro Proceso</h3>
+          <div>
+            <img src="" alt="Flecha 1" />
+            <p>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Voluptatibus ullam, nostrum aperiam, voluptatem, quas soluta quos aliquam natus quo accusamus sed doloremque vel beatae. Vitae omnis minima est praesentium corrupti.</p>
+            <img src="" alt="Flecha 2" />
+          </div>
+    </section>
+  )
+}
+
+export default Process
