+++ src/components/Navbar/Navbar.jsx
@@ -1,27 +1,25 @@
 import React from 'react'
 import './Navbar.css'
+import Instagram from '../../resources/icons/Instagram.svg'
+import Youtube from '../../resources/icons/Youtube.svg'
+import Ex from '../../resources/icons/X.svg'
 function Navbar() {
   return (
     <nav>
-        <img src="" alt="" />
-        <ul>
-          <li>1</li>
-          <li>2</li>
-          <li>3</li>
-          <li>4</li>
-        </ul>
-        <div>
-          <a href="">
-            <img src="" alt="" />
-          </a>
-          <a href="">
-            <img src="" alt="" />
-          </a>
-          <a href="">
-            <img src="" alt="" />
-          </a>
-        </div>
-      </nav>
+      {/* Opciones */}
+      <ul className='Opciones'>
+          <li><a href="HomePage">Home</a></li>
+          <li><a href="Recetario">Recetario</a></li>
+            <li><a href="Pasteles">Pasteles</a></li>
+          <li><a href="Nosotros">Nosotros</a></li>
+      </ul>
+      {/*Redes*/}
+      <ul className='Redes'>
+        <li className='Red'><a href="#"><img src={Instagram} alt="Instagram" /></a></li>
+        <li className='Red'><a href="#"><img src={Youtube} alt="Youtube" /></a></li>
+        <li className='Red'><a href="#"><img src={Ex} alt="Tiktok" /></a></li>
+      </ul>
+    </nav>
   )
 }
